refact: NutritionPage
