+++ src/components/NavigationBar.css
@@ -1,66 +1,66 @@
 /* 네비게이션 바 */
 .navigation-bar {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    position: fixed;
-    bottom: 0;
-    width: 100%;
-    height: 80px; /* 네비게이션 바 높이 */
-    background-color: #ffffff;
-    border-top: 1px solid #ddd;
-    z-index: 1000;
-  }
-  
-  /* 왼쪽과 오른쪽 섹션 */
-  .nav-left,
-  .nav-right {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    flex: 1;
-    padding: 0 10px;
-  }
-  
-  /* 중앙 섹션 */
-  .nav-center {
-    position: relative;
-    flex: 0;
-  }
-  
-  /* 네비게이션 아이템 */
-  .nav-item {
-    display: flex;
-    flex-direction: column; /* 아이콘 위 텍스트 아래 */
-    align-items: center;
-    justify-content: center;
-    width: 60px; /* 버튼 크기 */
-    height: 60px;
-    background-color: #f5f5f5;
-    border-radius: 50%;
-    margin: 0 5px;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-    transition: transform 0.2s ease;
-  }
-  
-  .nav-item:hover {
-    transform: scale(1.1);
-  }
-  
-  /* 네비게이션 아이콘 */
-  .nav-icon {
-    width: 30px;
-    height: 30px;
-    object-fit: contain;
-  }
-  
-  /* 네비게이션 텍스트 (작게 표시) */
-  .nav-label {
-    font-size: 10px; /* 텍스트 크기 */
-    color: #555; /* 텍스트 색상 */
-    margin-top: 5px; /* 아이콘과 텍스트 간격 */
-  }
-  
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  position: fixed;
+  bottom: 0;
+  width: 100%;
+  height: 80px; /* 네비게이션 바 높이 */
+  background-color: #ffffff;
+  border-top: 1px solid #ddd;
+  z-index: 1000;
+}
+
+/* 왼쪽과 오른쪽 섹션 */
+.nav-left,
+.nav-right {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  flex: 1;
+  padding: 0 10px;
+}
+
+/* 중앙 섹션 */
+.nav-center {
+  position: relative;
+  flex: 0;
+}
+
+/* 네비게이션 아이템 */
+.nav-item {
+  display: flex;
+  flex-direction: column; /* 아이콘 위 텍스트 아래 */
+  align-items: center;
+  justify-content: center;
+  width: 60px; /* 버튼 크기 */
+  height: 60px;
+  background-color: #f5f5f5;
+  border-radius: 50%;
+  margin: 0 5px;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+  transition: transform 0.2s ease;
+}
+
+.nav-item:hover {
+  transform: scale(1.1);
+}
+
+/* 네비게이션 아이콘 */
+.nav-icon {
+  width: 30px;
+  height: 30px;
+  object-fit: contain;
+}
+
+/* 네비게이션 텍스트 (작게 표시) */
+.nav-label {
+  font-size: 10px; /* 텍스트 크기 */
+  color: #555; /* 텍스트 색상 */
+  margin-top: 5px; /* 아이콘과 텍스트 간격 */
+}
+
 .camera-button {
   width: 80px;
   height: 80px;
@@ -81,6 +81,33 @@
   object-fit: contain;
 }
 
-  .camera-button:hover {
-    transform: translateY(-25px) scale(1.1);
-  }
+.camera-button:hover {
+  transform: translateY(-25px) scale(1.1);
+}
+
+/* 네비게이터 바 */
+.navigator {
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+  width: 100%;
+  height: 60px;
+  background-color: #ffffff;
+  position: fixed;
+  bottom: 0;
+  box-shadow: 0px -2px 6px rgba(0, 0, 0, 0.1);
+}
+
+.navigator-item {
+  text-align: center;
+  flex-grow: 1;
+  font-size: 16px;
+  color: #666666;
+  text-decoration: none;
+  cursor: pointer;
+}
+
+.navigator-item.active {
+  color: #00c5a1;
+  font-weight: bold;
+}
