+++ src/components/NutritionCarmera.css
@@ -1,0 +1,18 @@
+.camera {
+  background-color: #00c5a1;
+  color: white;
+  border: none;
+  border-radius: 50%;
+  width: 10px;
+  height: 80px;
+  font-size: 20px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  transform: translateY(-10px);
+  cursor: pointer;
+}
+
+.camera:hover {
+  background-color: #00c5a1;
+}
