+++ src/components/NutritionCarmera.js
@@ -1,4 +1,5 @@
 import React, { useRef } from 'react';
+import './NutritionCarmera.css';
 
 const NutritionCarmera = ({ setLoading }) => {
   const fileInputRef = useRef(null);
