+++ src/components/NutritionInfo.css
@@ -1,0 +1,14 @@
+.nutrition-info {
+  width: 90%;
+  background-color: #00c5a1;
+  padding: 15px;
+  border-radius: 8px;
+  box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.1);
+  margin: 20px 0;
+  color: white;
+}
+
+.nutrition-info p {
+  margin: 5px 0;
+  font-size: 14px;
+}
