+++ src/components/NutritionMeals.css
@@ -1,0 +1,39 @@
+/* 식사 섹터 */
+.meal-section {
+  width: 100%;
+}
+
+.meal-time {
+  background-color: #ffffff;
+  border-radius: 8px;
+  padding: 15px;
+  margin-bottom: 13px;
+  border-color: #00c5a1;
+  border-width: thick;
+  color: #00c5a1;
+
+  box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.1);
+}
+
+.meal-item {
+  padding: 5px;
+  margin: 5px 0;
+  background-color: #ffffff;
+  border-radius: 3px;
+  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
+  word-break: break-word; /* 음식 이름이 길어질 경우 줄바꿈 */
+}
+.meal-box {
+  padding: 10px;
+  background-color: #f9f9f9;
+  border-radius: 5px;
+  border: 1px solid #ddd;
+  min-height: 50px; /* 초기 높이 */
+  margin-top: -10px; /* 제목과 박스 간격 제거 */
+}
+
+.meal-container:not(:last-child) {
+  margin-bottom: 20px;
+  border-bottom: 1px solid #ddd;
+  padding-bottom: 20px;
+}
