+++ src/pages/NutritionPage.css
@@ -1,164 +1,19 @@
 body {
-    margin: 0;
-    padding: 0;
-    font-family: 'pretendard';
-    background-color: #f5f5f5;
-  }
-  
-  /* 메인 컨테이너 */
-  .nutrition-container {
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-    align-items: center;
-    height: 100vh;
-    max-width: 390px; /* iPhone 12 Pro 가로 길이 */
-    margin: 0 auto;
-    padding: 20px;
-    box-sizing: border-box;
-  }
-  
-  /* 달력 버튼 */
-  .calendar-section {
-    position: absolute; /* 부모 컨테이너 기준 위치 설정 */
-    top: 1px; /* 화면 상단 여백 */
-    right: 10px; /* 화면 오른쪽 여백 */
-    display: flex; /* 버튼과 날짜를 가로로 배치 */
-    align-items: center; /* 수직 가운데 정렬 */
-    gap: 8px; /* 버튼과 날짜 사이의 간격 */
-    white-space: nowrap;
-  }
-  .selected-date {
-    font-size: 16px; /* 폰트 크기 조정 */
-    font-weight: 600; /* 텍스트 굵기 */
-    width: 84px;
-    font-size: 16px;
-    color: #141414;
-    text-align: left;
-    
+  margin: 0;
+  padding: 0;
+  font-family: 'pretendard';
+  background-color: #f5f5f5;
+}
 
-   
-  }
-  /*캘린더 버튼*/
-  .calendar-button {
-    background-color: #f5f5f5;
-    color: #00c5a1;
-    border-color: #00c5a1;
-    border-radius: 30px;
-    padding: 10px 15px;
-    font-size: 12px;
-    cursor: pointer;
-  }
-  
-  .calendar-button:hover {
-    background-color: #00c5a1;
-  }
-  
-  /* 영양 정보 섹터 */
-  .nutrition-info {
-    width: 90%;
-    background-color: #00c5a1;
-    padding: 15px;
-    border-radius: 8px;
-    box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.1);
-    margin: 20px 0;
-    color : white;
-    
-    
-  }
-  
-  .nutrition-info p {
-    margin: 5px 0;
-    font-size: 14px;
-  }
-  
-  /* 식사 섹터 */
-  .meal-section {
-    width: 100%;
-    
-  }
-  
-  .meal-time {
-    background-color: #ffffff;
-    border-radius: 8px;
-    padding: 15px;
-    margin-bottom: 13px;
-    border-color: #00c5a1;
-    border-width: thick;
-    color : #00c5a1;
-   
-    box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.1);
-  }
-  
-  .meal-item {
-    padding: 5px;
-    margin: 5px 0;
-    background-color: #ffffff;
-    border-radius: 3px;
-    box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
-    word-break: break-word; /* 음식 이름이 길어질 경우 줄바꿈 */
-  }
-  .meal-box {
-    padding: 10px;
-    background-color: #f9f9f9;
-    border-radius: 5px;
-    border: 1px solid #ddd;
-    min-height: 50px; /* 초기 높이 */
-    margin-top: -10px; /* 제목과 박스 간격 제거 */
-  }
-  
-  .meal-container:not(:last-child) {
-    margin-bottom: 20px;
-    border-bottom: 1px solid #ddd;
-    padding-bottom: 20px;
-  }
-  
-  
-
-  
-  /* 네비게이터 바 */
-  .navigator {
-    display: flex;
-    justify-content: space-around;
-    align-items: center;
-    width: 100%;
-    height: 60px;
-    background-color: #ffffff;
-    position: fixed;
-    bottom: 0;
-    box-shadow: 0px -2px 6px rgba(0, 0, 0, 0.1);
-  }
-  
-  .navigator-item {
-    text-align: center;
-    flex-grow: 1;
-    font-size: 16px;
-    color: #666666;
-    text-decoration: none;
-    cursor: pointer;
-  }
-  
-  .navigator-item.active {
-    color: #00C5A1;
-    font-weight: bold;
-  }
-  
-  .camera {
-    background-color: #00C5A1;
-    color: white;
-    border: none;
-    border-radius: 50%;
-    width: 10px;
-    height: 80px;
-    font-size: 20px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    transform: translateY(-10px);
-    cursor: pointer;
-  }
-  
-  .camera:hover {
-    background-color: #00C5A1;
-  }
-  
+/* 메인 컨테이너 */
+.nutrition-container {
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  align-items: center;
+  height: 100vh;
+  max-width: 390px; /* iPhone 12 Pro 가로 길이 */
+  margin: 0 auto;
+  padding: 20px;
+  box-sizing: border-box;
+}
